fixed the rounding issues by using BigDecimal

+++ src/main/scala/com/world/wayne/menu/Items.scala
@@ -5,7 +5,7 @@ trait Items {
 
   def burgerType: String
 
-  def unitPrice: Double
+  def unitPrice: BigDecimal
 
-  def total: Double
+  def total: BigDecimal
 }
